Improved: Scan Speed for x86 by ~16%

+++ Reloaded.Memory.Sigscan/Instructions/GenericInstruction.cs
@@ -13,19 +13,19 @@ public struct GenericInstruction
     /// <summary>
     /// The value to match.
     /// </summary>
-    public ulong LongValue;
+    public nuint LongValue;
 
     /// <summary>
     /// The mask to apply before comparing with the value.
     /// </summary>
-    public ulong Mask;
+    public nuint Mask;
 
     /// <summary>
     /// Creates an instruction to match an 8 byte masked value at a given address.
     /// </summary>
     /// <param name="longValue">The value to be matched.</param>
     /// <param name="mask">The mask to match.</param>
-    public GenericInstruction(ulong longValue, ulong mask)
+    public GenericInstruction(nuint longValue, nuint mask)
     {
         LongValue   = longValue;
         Mask        = mask;
